+++ src/components/PreferencesForm.js
@@ -1,0 +1,14 @@
+import React from "react";
+import Typography from '@material-ui/core/Typography';
+import Link from '@material-ui/core/Link';
+import { makeStyles } from '@material-ui/core/styles';
+import { useForm } from "react-hook-form";
+
+
+export default function PreferencesForm() {
+
+    return (
+
+        <div>Preferences form</div>
+    );
+};
